+++ utils/index.ts
@@ -1,6 +1,4 @@
-
-export * from "./src/utils";
-export * from "./src/load";
-export * from "./src/color";
-export { default as EventEmitter } from './src/event-emitter';
-
+export * from './src/utils'
+export * from './src/load'
+export * from './src/color'
+export { default as EventEmitter } from './src/event-emitter'