+++ vue-components/bin/gen-tags.json.cjs
@@ -1,22 +1,20 @@
-const fs =  require('node:fs');
-const webJson = require("../web-types.json")
+const fs = require('node:fs')
+const webJson = require('../web-types.json')
 
 // 这里替换成你的 vue-components 数组
-const vueComponents = webJson.contributions.html["vue-components"]
+const vueComponents = webJson.contributions.html['vue-components']
 
-const tagsJson = {};
-
-
+const tagsJson = {}
 
 for (const comp of vueComponents) {
-    const props = comp.props?.map(p => p.name) || [];
-    const events = comp.js?.events?.map(e => e.name) || [];
+    const props = comp.props?.map((p) => p.name) || []
+    const events = comp.js?.events?.map((e) => e.name) || []
 
     tagsJson[comp.name] = {
         attributes: [...props, ...events],
         description: comp.description || ''
-    };
+    }
 }
 
-fs.writeFileSync('tags.json', JSON.stringify(tagsJson, null, 2), 'utf-8');
-console.log('tags.json 生成完成');
+fs.writeFileSync('tags.json', JSON.stringify(tagsJson, null, 2), 'utf-8')
+console.log('tags.json 生成完成')