+++ vue-components/src/FormTable/index.ts
@@ -1,16 +1,11 @@
 import { withInstall } from '../install'
 import type { SFCWithInstall } from '../component-install-types'
-import ZJFormTableContainer from './Container.ts';
-import ZJTableEditColumn from './EditColumn.vue';
-import ZJTableEditColumnAction from './EditColumnAction.vue';
-import ZJFormTableSearchItem from './SearchItem.vue';
-import ZJFormTableItem from './Item.vue';
-export {
-    ZJTableEditColumn as TableEditColumn,
-    ZJTableEditColumnAction as TableEditColumnAction,
-    ZJFormTableSearchItem as FormTableSearchItem,
-    ZJFormTableItem as FormTableItem,
-}
+import ZJFormTableContainer from './Container.ts'
+import ZJTableEditColumn from './EditColumn.vue'
+import ZJTableEditColumnAction from './EditColumnAction.vue'
+import ZJFormTableSearchItem from './SearchItem.vue'
+import ZJFormTableItem from './Item.vue'
+export { ZJTableEditColumn as TableEditColumn, ZJTableEditColumnAction as TableEditColumnAction, ZJFormTableSearchItem as FormTableSearchItem, ZJFormTableItem as FormTableItem }
 export const FormTableContainer: SFCWithInstall<typeof ZJFormTableContainer> = withInstall(ZJFormTableContainer, {
     TableEditColumn: ZJTableEditColumn,
     TableEditColumnAction: ZJTableEditColumnAction,
@@ -22,4 +17,3 @@ export default FormTableContainer
 
 export * from './types'
 export type FormTableContainerInstance = InstanceType<typeof ZJFormTableContainer> & unknown
-