+++ vue-hooks/vite.config.ts
@@ -1,4 +1,4 @@
-import {defineConfig} from "vite";
+import { defineConfig } from 'vite'
 import { dirname, resolve } from 'node:path'
 import { fileURLToPath } from 'node:url'
 const __dirname = dirname(fileURLToPath(import.meta.url))
@@ -6,14 +6,14 @@ export default defineConfig({
     build: {
         lib: {
             entry: {
-                index: resolve(__dirname, 'index.ts'),
+                index: resolve(__dirname, 'index.ts')
             },
             name: '$ZJHooks',
             fileName: (format, entryName) => {
                 if (format === 'cjs') return `${entryName}.${format}`
                 return `${entryName}.${format}.js`
             },
-            formats: ["es", "cjs", "umd"],
+            formats: ['es', 'cjs', 'umd']
         },
         rollupOptions: {
             external: ['vue', '@zj-library/utils', '@zj-library/vue-components', '@zj-library/vue-hooks'],
@@ -23,13 +23,9 @@ export default defineConfig({
                     '@zj-library/vue-components': '$ZJVueComponents',
                     '@zj-library/utils': '$ZJUtils',
                     '@zj-library/vue-hooks': '$ZJHooks',
-                    vue: 'Vue',
-                },
-            },
-        },
-    },
+                    vue: 'Vue'
+                }
+            }
+        }
+    }
 })
-
-
-
-