+++ utils/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ### Patch Changes
 
-- f909dcc: 初次提交
+-   f909dcc: 初次提交